requirements.txt
numpy==1.23.2
pandas==2.2.3
unstructured-ingest==0.3.12
unstructured-ingest[remote]
unstructured[all-docs]==0.13.7
python-dotenv==1.0.1
langchain-core==0.3.31
langchain-openai==0.3.0
streamlit==1.41.1
torch==2.5.1
langchain==0.3.15
chromadb==0.6.3
langchain-community==0.3.15
boto3
botocore
ipython

-e .